# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Swagger documentation for the OpenCost costmodel HTTP API.",
    "title": "OpenCost API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/kapis/costwise.wiztelemetry.io/v1alpha1/allocation/autocomplete": {
      "get": {
        "description": "兼容 Kubecost 的 allocation autocomplete 接口。当前支持 label、annotation、cluster、node、namespace、pod、container、controller、controllerKind、providerID、service。",
        "tags": [
          "Allocation"
        ],
        "summary": "查询分配字段自动补全候选项",
        "parameters": [
          {
            "description": "时间窗口。必填。",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "字段名。支持 label、annotation、cluster、node、namespace、pod、container、controller、controllerKind、providerID、service。",
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "搜索关键字，按包含关系过滤候选项。",
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "分配过滤条件。",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
...
    "/kapis/costwise.wiztelemetry.io/v1alpha1/efficiency": {
      "get": {
        "description": "基于请求窗口和聚合维度计算资源效率、推荐请求值和潜在成本节省，支持 CSV 导出。",
        "tags": [
          "Efficiency"
        ],
        "summary": "查询效率优化建议",
        "parameters": [
          {
            "description": "时间窗口。必填。",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "聚合维度。默认 pod。",
            "name": "aggregate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "分配过滤条件。",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "推荐资源计算缓冲系数。默认 1.2。",
            "name": "bufferMultiplier",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "返回格式；传 csv 导出 CSV。",
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "costmodel.Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "data": {},
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "warning": {
            "type": "string"
          }
        }
      }
    }
  }
}